refactor(infermap): single-source string-distance on goldenmatch's kernel fallback
Single-kernel-collapse R5 consolidation, package 1 of 3. The
`ts-no-duplicate-kernel-math` gate drops from 5 advisory findings to 2.

infermap vendored its own jaro / jaro-winkler / levenshtein. They now alias
goldenmatch's, which are the parity-gated pure-TS fallback for the
goldenmatch-score-core Rust kernel (floating-point-identical to rapidfuzz at
5.6e-17 over a 2005-pair sweep). This also closes a Python<->TS parity gap:
infermap's Python `fuzzy_name` already reuses
`goldenmatch-score-core::jaro_winkler_similarity`, so the TS fork was the only
surface still running its own math.

The fork had drifted, exactly as the gate predicts. It carried all three
pre-#879 bugs goldenmatch had already fixed:

  * unfloored `t/2` transposition instead of floored — jaroWinkler("saturday",
    "sunday") 0.7475 -> 0.7775, the exact pair and values #879 cites
  * Winkler boost at `jaro >= 0.7` instead of rapidfuzz's strict `> 0.7`
  * UTF-16 code-unit iteration instead of codepoints

Measured old-vs-new over 75.7K pairs: 57 jaro/JW and 62 levenshtein
disagreements, all non-BMP. A 240K-pair ASCII-only sweep found zero, so ASCII
column names -- infermap's normal input -- are unaffected apart from the
transposition class.

Mechanism: `core/scorer.ts` is 1794 lines and transitively pulls the
reference-data tables and WASM registry, so importing it wholesale is a
disproportionate payload for three functions. Extracted the primitives into a
zero-import leaf `core/stringDistance.ts` (moved verbatim), published as the
`goldenmatch/core/string-distance` subpath. `core/scorer.ts` re-exports them, so
no existing import and no public API changed. infermap keeps goldenmatch as a
devDependency and tsup inlines the leaf, the established
`goldenmatch-wasm-runtime` pattern -- no published runtime dep is added.

Also drops the vendored `prefixScale` third argument (no call site ever passed
it; a tunable prefix scale is by definition not kernel-parity behaviour).

Verified: goldenmatch 3671 tests + infermap 279 tests green, both typecheck and
build clean, published .d.ts carries self-contained declarations with no
unresolvable `goldenmatch/...` type reference.
