+++ .ast-grep/README.md
@@ -48,7 +48,7 @@ advisory (reported, non-blocking) so new rules can land gradually.
 |---|---|---|
 | `ts-no-empty-catch` | error | empty `catch {}` silently swallows errors — log or re-throw |
 | `ts-no-spread-math-min-max` | error | `Math.min/max(...array)` throws on >65K elements — the 11 real sites were fixed (this PR), so it's now a hard guard against reintroduction |
-| `ts-no-duplicate-kernel-math` | warning | a hand-rolled `levenshtein`/`jaro`/`soundex`/… outside the sanctioned pure-TS fallback modules is a second copy that drifts from the Rust `*-core` kernel and the Python port — import the existing export ([single-kernel-collapse R5](../context-network/architecture/single-kernel-collapse-roadmap.md)). The parity-gated fallbacks (`goldenmatch` `core/scorer.ts`, `core/transforms.ts`, `core/perceptualWasm.ts`) are ignored — they are the permanent edge-safe fallback surface, not drift. Currently **5 advisory findings** (goldencheck `fuzzy-values.ts`, goldenflow `phonetic.ts`, infermap `string-distance.ts` ×3) = the consolidation backlog; promote to `error` once they're gone |
+| `ts-no-duplicate-kernel-math` | warning | a hand-rolled `levenshtein`/`jaro`/`soundex`/… outside the sanctioned pure-TS fallback modules is a second copy that drifts from the Rust `*-core` kernel and the Python port — import the existing export ([single-kernel-collapse R5](../context-network/architecture/single-kernel-collapse-roadmap.md)). The parity-gated fallbacks (`goldenmatch` `core/stringDistance.ts`, `core/scorer.ts`, `core/transforms.ts`, `core/perceptualWasm.ts`) are ignored — they are the permanent edge-safe fallback surface, not drift. Currently **2 advisory findings** (goldencheck `fuzzy-values.ts`, goldenflow `phonetic.ts`) = the remaining consolidation backlog; infermap `string-distance.ts` (×3) was consolidated onto `goldenmatch/core/string-distance`. Promote to `error` once the last two are gone |
 
 **Rust kernels** (`packages/rust/extensions/`):
