+++ packages/typescript/infermap/tsup.config.ts
@@ -9,7 +9,10 @@ export default defineConfig({
     cli: "src/cli.ts",
   },
   format: ["esm", "cjs"],
-  dts: { resolve: ["goldenmatch-wasm-runtime"] },
+  // `resolve` must name the PACKAGE, not the subpath specifier — it rolls the
+  // referenced .d.ts into ours so the published types don't point at
+  // `goldenmatch`, which is only a devDependency here.
+  dts: { resolve: ["goldenmatch-wasm-runtime", "goldenmatch"] },
   sourcemap: true,
   clean: true,
   target: "node20",
@@ -21,8 +24,12 @@ export default defineConfig({
   loader: { ".wasm": "copy" },
   publicDir: false,
   onSuccess: "node scripts/copy_wasm_artifact.mjs",
-  // Inline the tiny WASM plumbing so it's not a published runtime dep.
-  noExternal: ["goldenmatch-wasm-runtime"],
+  // Inline the tiny WASM plumbing so it's not a published runtime dep. Same for
+  // the goldenmatch edit-distance leaf (`goldenmatch/core/string-distance`):
+  // infermap single-sources those primitives rather than vendoring a second
+  // copy, but goldenmatch stays a devDependency — the leaf has zero imports, so
+  // inlining it costs ~150 lines, not the goldenmatch package.
+  noExternal: ["goldenmatch-wasm-runtime", "goldenmatch"],
   external: [
     // Runtime-only wasm-bindgen glue (dynamic-imported in enableInfermapWasm);
     // absent in a default checkout. Mark external so esbuild never resolves it.
